+++ src/main/java/com/kazh_kvetk/fuzzy_inference/machine/Machine.java
@@ -5,7 +5,7 @@ import com.kazh_kvetk.fuzzy_inference.machine.units.Rule;
 import java.util.Queue;
 
 public interface Machine {
-  void accumulate(Rule rule, Queue<Integer> inputs);
-  Integer mostConsistentRuleResult();
+  void accumulate(Rule rule, Queue<Long> inputs);
+  Long mostConsistentRuleResult();
   ConclusionFunction mostConsistentFunction();
 }
